requirements.txt

flask==2.3.3
flask-socketio==5.3.6
flask-cors==4.0.0
python-dotenv==1.0.0
langchain==0.1.5
langchain-openai==0.1.5
langchainhub==0.1.14
langgraph==0.0.29
